% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to.R
\name{clean_r}
\alias{clean_r}
\title{Clean R file}
\usage{
clean_r(rfile)
}
\arguments{
\item{rfile}{character l'adresse du fichier a nettoyer par exemple : 
"inst/rstudio/templates/project/ressources/m7/corrections/exo2.R"}
}
\description{
Nettoyage des scripts R préparés par la fonction `rmd_to()`
}
