% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{exo_modules_skeleton}
\alias{exo_modules_skeleton}
\title{Function to define the project exo_modules}
\usage{
exo_modules_skeleton(path, ...)
}
\arguments{
\item{path}{path where the project is create}

\item{...}{params define by the user a the project creation, dans notre cas module suivi par le stagiaire 
(cf liste des parametres possible dans le fichier. dcf de `inst/rstudio/templates/project`)}
}
\value{
create a project
}
\description{
Function to define the project exo_modules
}
\keyword{internal}
