% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_r_files.R
\name{extract_r_files}
\alias{extract_r_files}
\title{Liste les rmd dans le dossiers inst/vignettes du module et les place dans les dossiers t Extrait les enonces et les corrections des rmd et les place dans les dossiers}
\usage{
extract_r_files(module = 5)
}
\arguments{
\item{module}{le numero du dossier}
}
\value{
nothing
}
\description{
Liste les rmd dans le dossiers inst/vignettes du module et les place dans les dossiers t Extrait les enonces et les corrections des rmd et les place dans les dossiers
}
\examples{
\dontrun{
extract_r_files(module = 5)
}
}
