% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{module_data}
\alias{module_data}
\title{get extdata file needed for a module}
\usage{
module_data(m)
}
\arguments{
\item{m}{module number}
}
\value{
a list of files
}
\description{
get extdata file needed for a module
}
\keyword{internal}
