% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{pkg_file}
\alias{pkg_file}
\title{get files from the exercices package}
\usage{
pkg_file(...)
}
\arguments{
\item{...}{other params from the systeme.file fonction}
}
\description{
get files from the exercices package
}
