% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to.R
\name{rmd_to_enonce}
\alias{rmd_to_enonce}
\title{rmd_to_enonce}
\usage{
rmd_to_enonce(module, rmdfile)
}
\arguments{
\item{module}{"m1"}

\item{rmdfile}{"exo1.rmd"}
}
\description{
Mise en place d'un script de consignes à partir de l'exercice Rmd saisi dans `inst/vignettes`
}
