% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_exo.R
\name{charge_exo}
\alias{charge_exo}
\title{Charger les exos des modules}
\usage{
charge_exo(...)
}
\arguments{
\item{...}{chemin partant de vignettes vers le l'exercice}
}
\value{

}
\description{
Charger les exos des modules
}
\examples{
charge_exo("m5", "exo1.rmd")
}
