% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_r_files.R
\name{extract_r_files}
\alias{extract_r_files}
\title{Liste les exercices d'un module au format rmd, presents dans le dossiers inst/vignettes du module, puis en extrait/nettoie/colle le contenu  dans des scripts R d'enonces et de corrections, rendus accessibles aux stagiaires utilisant le package savoirFr pour réaliser les exercices, au niveau du dossier inst/rstudio/templates/project/ressources du module. N'impacte que le package savoirfR.}
\usage{
extract_r_files(module = 5)
}
\arguments{
\item{module}{le numero du dossier}
}
\value{
nothing
}
\description{
Liste les exercices d'un module au format rmd, presents dans le dossiers inst/vignettes du module, puis en extrait/nettoie/colle le contenu  dans des scripts R d'enonces et de corrections, rendus accessibles aux stagiaires utilisant le package savoirFr pour réaliser les exercices, au niveau du dossier inst/rstudio/templates/project/ressources du module. N'impacte que le package savoirfR.
}
\examples{
\dontrun{
extract_r_files(module = 5)
}
}
