% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to.R
\name{rmd_to}
\alias{rmd_to}
\title{rmd_to}
\usage{
rmd_to(module, rmdfile, folder = "corrections", purl = TRUE)
}
\arguments{
\item{module}{"m1"}

\item{rmdfile}{"exo1.rmd"}

\item{folder}{corrections ou enonces}

\item{purl}{TRUE (pour extraire le code R) ou FALSE pour ignorer le code R}
}
\value{

}
\description{
rmd_to
}
