% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to.R
\name{rmd_to}
\alias{rmd_to}
\title{Title}
\usage{
rmd_to(module, rmdfile, folder = "corrections", purl = TRUE)
}
\arguments{
\item{module}{}

\item{rmdfile}{}

\item{folder}{}

\item{purl}{}
}
\value{

}
\description{
Title
}
