% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to.R
\name{rmd_to_correction}
\alias{rmd_to_correction}
\title{rmd_to_correction}
\usage{
rmd_to_correction(module, rmdfile)
}
\arguments{
\item{module}{}

\item{rmdfile}{}
}
\value{

}
\description{
rmd_to_correction
}
